requirements.txt
requests>=2.31.0
pytest>=7.0.0
pytest-cov>=4.0.0
playwright>=1.40.0
pytest-playwright>=0.4.0
python-dotenv>=1.0.0
